Make the build reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
fluids could not be built reproducibly.

This is due to the documentation automatically including the current
year in a copyright footer, which means that the build will change
depending on when you build.

This patch uses SOURCE_DATE_EPOCH [1] to make this "build date"
configurable/pluggable.

I originally filed this in Debian as bug #1004391 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/1004391